+++ CHANGELOG.md
@@ -25,6 +25,7 @@ IMPROVEMENTS:
 
 BUG FIXES:
 
+* Add the missing single-block constraint to the Plugin Framework schemas for all `auth_login_*` provider configuration blocks.
 * Fixed the token namespace being set as the provider namespace, even when `set_namespace_from_token` was `false`. ([#2926](https://github.com/hashicorp/terraform-provider-vault/pull/2926/))
 * `vault_pki_secret_backend_role`: Fix crash when the Vault client was not successfully initialized ([#2801](https://github.com/hashicorp/terraform-provider-vault/pull/2801))